ctdb/server: cppcheck: fix shiftTooManyBitsSigned error
Fixes

ctdb/server/ipalloc_lcp2.c:61: error: shiftTooManyBitsSigned: Shifting signed 32-bit value by 31 bits is undefined behaviour <--[cppcheck]

+++ ctdb/server/ipalloc_lcp2.c
@@ -58,7 +58,7 @@ static uint32_t ip_distance(ctdb_sock_addr *ip1, ctdb_sock_addr *ip2)
 			/* Count number of leading zeroes.
 			 * FIXME? This could be optimised...
 			 */
-			while ((x & (1 << 31)) == 0) {
+			while ((x & ((uint32_t)1 << 31)) == 0) {
 				x <<= 1;
 				distance += 1;
 			}
